+++ handler/handler_test.go
@@ -117,7 +117,7 @@ func TestPush(t *testing.T) {
 			httprouter.Param{Key: "instance", Value: "testinstance"},
 		},
 	)
-	if expected, got := http.StatusInternalServerError, w.Code; expected != got {
+	if expected, got := http.StatusBadRequest, w.Code; expected != got {
 		t.Errorf("Wanted status code %v, got %v.", expected, got)
 	}
 	if !mms.lastWriteRequest.Timestamp.IsZero() {
@@ -159,6 +159,9 @@ func TestPush(t *testing.T) {
 	if expected, got := `name:"another_metric" type:UNTYPED metric:<label:<name:"instance" value:"testinstance" > label:<name:"job" value:"testjob" > untyped:<value:42 > > `, mms.lastWriteRequest.MetricFamilies["another_metric"].String(); expected != got {
 		t.Errorf("Wanted metric family %v, got %v.", expected, got)
 	}
+	if _, ok := mms.lastWriteRequest.MetricFamilies["push_time_seconds"]; !ok {
+		t.Errorf("Wanted metric family push_time_seconds missing.")
+	}
 
 	// With job name and no instance name and text content.
 	mms.lastWriteRequest = storage.WriteRequest{}
@@ -195,6 +198,30 @@ func TestPush(t *testing.T) {
 		t.Errorf("Wanted metric family %v, got %v.", expected, got)
 	}
 
+	// With job name and instance name and timestamp specified.
+	mms.lastWriteRequest = storage.WriteRequest{}
+	req, err = http.NewRequest(
+		"POST", "http://example.org/",
+		bytes.NewBufferString("a 1\nb 1 1000\n"),
+	)
+	if err != nil {
+		t.Fatal(err)
+	}
+	w = httptest.NewRecorder()
+	handler(
+		w, req,
+		httprouter.Params{
+			httprouter.Param{Key: "job", Value: "testjob"},
+			httprouter.Param{Key: "labels", Value: "/instance/testinstance"},
+		},
+	)
+	if expected, got := http.StatusBadRequest, w.Code; expected != got {
+		t.Errorf("Wanted status code %v, got %v.", expected, got)
+	}
+	if !mms.lastWriteRequest.Timestamp.IsZero() {
+		t.Errorf("Write request timestamp unexpectedly set: %#v", mms.lastWriteRequest)
+	}
+
 	// With job name and instance name and text content, legacy handler.
 	mms.lastWriteRequest = storage.WriteRequest{}
 	req, err = http.NewRequest(
@@ -229,6 +256,33 @@ func TestPush(t *testing.T) {
 	}
 	if expected, got := `name:"another_metric" type:UNTYPED metric:<label:<name:"instance" value:"testinstance" > label:<name:"job" value:"testjob" > untyped:<value:42 > > `, mms.lastWriteRequest.MetricFamilies["another_metric"].String(); expected != got {
 		t.Errorf("Wanted metric family %v, got %v.", expected, got)
+	}
+	if _, ok := mms.lastWriteRequest.MetricFamilies["push_time_seconds"]; !ok {
+		t.Errorf("Wanted metric family push_time_seconds missing.")
+	}
+
+	// With job name and instance name and timestamp, legacy handler.
+	mms.lastWriteRequest = storage.WriteRequest{}
+	req, err = http.NewRequest(
+		"POST", "http://example.org/",
+		bytes.NewBufferString("a 1\nb 1 1000\n"),
+	)
+	if err != nil {
+		t.Fatal(err)
+	}
+	w = httptest.NewRecorder()
+	legacyHandler(
+		w, req,
+		httprouter.Params{
+			httprouter.Param{Key: "job", Value: "testjob"},
+			httprouter.Param{Key: "instance", Value: "testinstance"},
+		},
+	)
+	if expected, got := http.StatusBadRequest, w.Code; expected != got {
+		t.Errorf("Wanted status code %v, got %v.", expected, got)
+	}
+	if !mms.lastWriteRequest.Timestamp.IsZero() {
+		t.Errorf("Write request timestamp unexpectedly set: %#v", mms.lastWriteRequest)
 	}
 
 	// With job name and instance name and text content and job and instance labels.